Working in wrong repo. Pushing for ajax-patch.
Worked in ajax-delete on the solution branch, as it will soon be the starter
code for the ajax-patch repo, and this will be the solution branch for the
ajax-patch repo. Should've done this as a new branch (Solution) off of master
in the ajax-patch repo, but the starter code wasn't there yet. Will reference
this in the related PR for the master branch of the ajax-patch repo.